requirements.txt
crewai
python-dotenv
neo4j
yfinance
newsapi-python
sec-api
pandas
numpy
requests 